% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.plantstress_sri}
\alias{plot.plantstress_sri}
\title{Plot a Stress Signature}
\usage{
\method{plot}{plantstress_sri}(x, ...)
}
\arguments{
\item{x}{A \code{plantstress_sri} object.}

\item{...}{Passed to \code{\link[=plot_stress_signature]{plot_stress_signature()}}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Shortcut for \code{\link[=plot_stress_signature]{plot_stress_signature()}}, so that \code{plot()} does the expected
thing on the object returned by \code{\link[=calculate_sri]{calculate_sri()}}.
}
\examples{
data(brachiaria_stress)
sri <- calculate_sri(brachiaria_stress,
  treatment = "drought_level", control = "control",
  traits = c("Fv_Fm", "PIabs", "A", "gs", "RWC"), verbose = FALSE
)
plot(sri)
}
\seealso{
\code{\link[=plot_stress_signature]{plot_stress_signature()}}
}
